requirements.txt
google-cloud-logging>=3.1.0
requests>=2.28.0
functions-framework>=3.0.0
